+++ src/main/java/com/cdp/tdp/controller/UserController.java
@@ -1,8 +1,10 @@
 package com.cdp.tdp.controller;
 
+import com.cdp.tdp.domain.User;
 import com.cdp.tdp.dto.JwtResponse;
 import com.cdp.tdp.dto.SignupRequestDto;
 import com.cdp.tdp.dto.UserDto;
+import com.cdp.tdp.security.UserDetailsImpl;
 import com.cdp.tdp.service.UserService;
 import com.cdp.tdp.util.JwtTokenUtil;
 import lombok.RequiredArgsConstructor;
@@ -11,17 +13,53 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.DisabledException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.web.bind.annotation.*;
-import com.cdp.tdp.util.JwtTokenUtil;
+
+import java.sql.SQLException;
 
 @RequiredArgsConstructor
 @RestController
 public class UserController {
 
+    private final JwtTokenUtil jwtTokenUtil;
+    private final AuthenticationManager authenticationManager;
+    private final UserDetailsService userDetailsService;
+    private final UserService userService;
 
+
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ResponseEntity<?> createAuthenticationToken(@RequestBody UserDto userDto) throws Exception {
+        authenticate(userDto.getUsername(), userDto.getPassword());
+        final UserDetails userDetails = userDetailsService.loadUserByUsername(userDto.getUsername());
+        final String token = jwtTokenUtil.generateToken(userDetails);
+        return ResponseEntity.ok(new JwtResponse(token, userDetails.getUsername()));
+    }
+
+    @PostMapping(value = "/signup")
+    public ResponseEntity<?> createUser(@RequestBody SignupRequestDto userDto) throws Exception {
+        userService.registerUser(userDto);
+        authenticate(userDto.getUsername(), userDto.getPassword());
+        final UserDetails userDetails = userDetailsService.loadUserByUsername(userDto.getUsername());
+        final String token = jwtTokenUtil.generateToken(userDetails);
+        return ResponseEntity.ok(new JwtResponse(token, userDetails.getUsername()));
+    }
+
+    private void authenticate(String username, String password) throws Exception {
+        try {
+            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(username, password));
+        } catch (DisabledException e) {
+            throw new Exception("USER_DISABLED", e);
+        } catch (BadCredentialsException e) {
+            throw new Exception("INVALID_CREDENTIALS", e);
+        }
+    }
+
+    @GetMapping(value = "/user")
+    public User readUser(@AuthenticationPrincipal UserDetailsImpl userDetails) throws SQLException {
+        User user = (User) userDetails.getUser();
+        return user;
+    }
 }
